Update finalproj.py
Just added more parameters to the class Tweet which would make sure the global variables for pos_ws and neg_ws were saved as local variables in the class. That way, if the list of positive and negative words was called something else globally, it would still work within the class.